operator functions moved to trait
