+++ crates/operator/src/lib.rs
@@ -4,7 +4,5 @@
 pub mod client;
 /// Error
 pub mod error;
-/// Operator
-pub mod operator;
 /// Traits
 pub mod traits;
